e2e-tests: migrated to updated playwright-test runner
+ this also meant a migration to fixture based tests
+ getting rid of the workaround to ensure execution order

Changelog: None
