+++ .gitlab-ci.yml
@@ -33,16 +33,29 @@ test:lint:
     - npm ci --cache .npm --prefer-offline
     - npm run lint
 
-test:acceptance:
+.template:test:acceptance:
   stage: e2e-test
-  except:
-    - /^(staging|saas-[a-zA-Z0-9.]+)$/
   image: tiangolo/docker-with-compose
-  dependencies:
-    - build:docker
-    - test:prep
+  needs:
+    - job: build:docker
+      artifacts: true
   services:
     - docker:19.03.5-dind
+  artifacts:
+    expire_in: 2w
+    paths:
+      - screenshots
+      - logs
+      - junit
+    when: on_failure
+    reports:
+      junit:
+        - junit/results.xml
+
+test:acceptance:
+  extends: .template:test:acceptance
+  except:
+    - /^(staging|saas-[a-zA-Z0-9.]+)$/
   before_script:
     - export DOCKER_TAG=${CI_COMMIT_REF_SLUG:-master}
     - export SERVICE_IMAGE=${DOCKER_REPOSITORY}:${DOCKER_TAG}
@@ -51,26 +64,14 @@ test:acceptance:
   script:
     - docker load -i image.tar
     - docker tag $SERVICE_IMAGE $DOCKER_REPOSITORY:pr
-    - git clone --single-branch https://github.com/mendersoftware/integration.git
+    - git clone --no-tags --depth=1 --single-branch https://github.com/mendersoftware/integration.git
     - GUI_REPOSITORY=$(pwd) INTEGRATION_PATH=$(pwd)/integration ./tests/e2e_tests/run
     - docker rmi $DOCKER_REPOSITORY:pr
-  artifacts:
-    expire_in: 2w
-    paths:
-      - screenshots
-      - logs
-    when: on_failure
 
 test:acceptance:enterprise:
-  stage: e2e-test
+  extends: .template:test:acceptance
   only:
     - master
-  image: tiangolo/docker-with-compose
-  dependencies:
-    - build:docker
-    - test:prep
-  services:
-    - docker:19.03.5-dind
   variables:
     TEST_ENVIRONMENT: enterprise
   before_script:
@@ -87,12 +88,6 @@ test:acceptance:enterprise:
     - git clone --no-tags --depth=1 --single-branch https://github.com/mendersoftware/integration.git
     - GUI_REPOSITORY=$(pwd) INTEGRATION_PATH=$(pwd)/integration ./tests/e2e_tests/run --enterprise
     - docker rmi $DOCKER_REPOSITORY:pr
-  artifacts:
-    expire_in: 2w
-    paths:
-      - screenshots
-      - logs
-    when: on_failure
 
 test:unit:
   stage: test
