+++ tests/e2e_tests/fixtures/fixtures.ts
@@ -1,43 +1,19 @@
-const { v4: uuid } = require('uuid');
-// import * as playwrightGlobalSetup from 'jest-playwright-preset/setup';
-// import { chromium } from 'playwright';
-
-// module.exports = async function globalSetup(globalConfig) {
-//   await playwrightGlobalSetup(globalConfig);
-
-//   let browser = await chromium.launch();
-//   let context = await browser.newContext({ ignoreHTTPSErrors: true });
-//   const page = await context.newPage();
-
-//   // your login function
-//   // await doLogin(page);
-
-//   // store authentication data
-//   const storage = await page.context().storageState();
-//   process.env.STORAGE = JSON.stringify(storage);
-// };
-
-import { folio as base } from '@playwright/test';
+import base, { Page } from '@playwright/test';
+import { v4 as uuid } from 'uuid';
+import { baseUrlToDomain, login } from '../utils/commands';
 
 type TestFixtures = {
   baseUrl: string;
   config: Object;
+  environment: string;
+  loggedInPage: Page;
   username: string;
   password: string;
   demoDeviceName: string;
 };
 
-const defaultConfig = {
-  baseUrl: 'https://docker.mender.io/',
-  username: 'mender-demo@example.com',
-  password: 'mysecretpassword!123',
-  demoDeviceName: 'release-v1'
-};
-
-const stagingConfig = {
-  username: `${uuid()}@example.com`,
-  password: uuid()
-};
+process.env.STAGING_USER = process.env.STAGING_USER ?? `${uuid()}@example.com`;
+process.env.STAGING_PASSWORD = process.env.STAGING_PASSWORD ?? uuid();
 
 const urls = {
   localhost: 'https://docker.mender.io/',
@@ -45,66 +21,55 @@ const urls = {
   production: 'https://hosted.mender.io/'
 };
 
-const fixtures = base.extend<{}, TestFixtures, { environment: string; usernameParam: string; passwordParam: string }>();
-fixtures.contextOptions.override(async ({ contextOptions }, runTest) => {
-  await runTest({
-    ...contextOptions,
-    acceptDownloads: true,
-    ignoreHTTPSErrors: true,
-    recordVideo: process.env.PWVIDEO ? { dir: 'screenshots' } : undefined,
-    viewport: { width: 1600, height: 900 }
-  });
+const defaultConfig = {
+  baseUrl: urls.localhost,
+  username: 'mender-demo@example.com',
+  password: 'mysecretpassword!123',
+  demoDeviceName: 'release-v1'
+};
+
+const test = base.extend<TestFixtures>({
+  loggedInPage: async ({ baseUrl, context, password, username }, use) => {
+    // const storageState = JSON.parse(process.env.STORAGE || '{}');
+    // let context: BrowserContext = await browser.newContext({ storageState });
+    await context.grantPermissions(['clipboard-read'], { origin: baseUrl });
+    const domain = baseUrlToDomain(baseUrl);
+    const { token, userId } = await login(username, password, baseUrl);
+    await context.addCookies([
+      { name: 'JWT', value: token, path: '/', domain },
+      { name: `${userId}-onboarded`, value: 'true', path: '/', domain },
+      { name: 'cookieconsent_status', value: 'allow', path: '/', domain }
+    ]);
+    const page = await context.newPage();
+    await page.goto(`${baseUrl}ui/`);
+    await page.evaluate(({ userId }) => localStorage.setItem(`${userId}-onboarding`, JSON.stringify({ complete: true })), { userId });
+    await page.waitForSelector('text=License information');
+    await use(page);
+    await context.storageState({ path: 'storage.json' });
+  },
+  environment: async ({}, use) => {
+    const environment = process.env.TEST_ENVIRONMENT ? process.env.TEST_ENVIRONMENT : 'localhost';
+    await use(environment);
+  },
+  username: async ({ environment }, use) => {
+    let username = defaultConfig.username;
+    if (environment === 'staging') {
+      username = process.env.STAGING_USER;
+    }
+    await use(username);
+  },
+  password: async ({ environment }, use) => {
+    let password = defaultConfig.password;
+    if (environment === 'staging') {
+      password = process.env.STAGING_PASSWORD;
+    }
+    await use(password);
+  },
+  baseUrl: async ({ environment }, use) => {
+    const baseUrl = urls[environment] || defaultConfig.baseUrl;
+    await use(baseUrl);
+  },
+  demoDeviceName: defaultConfig.demoDeviceName
 });
 
-fixtures.browserOptions.override(async ({ browserOptions }, runTest) => {
-  await runTest({
-    ...browserOptions, // Default options
-    args: ['--disable-dev-shm-usage', '--disable-web-security'],
-    slowMo: 100
-  });
-});
-
-fixtures.environment.initParameter('test environment', process.env.TEST_ENVIRONMENT ? process.env.TEST_ENVIRONMENT : 'localhost');
-fixtures.usernameParam.initParameter('username ', `${uuid()}@example.com`);
-fixtures.passwordParam.initParameter('password', uuid());
-
-fixtures.config.init(async ({}, run) => {
-  await run(defaultConfig);
-}, {});
-
-fixtures.baseUrl.init(async ({ environment }, run) => {
-  const baseUrl = urls[environment] || defaultConfig.baseUrl;
-  await run(baseUrl);
-}, {});
-
-fixtures.username.init(async ({ environment, usernameParam }, run) => {
-  let username = defaultConfig.username;
-  if (environment === 'staging') {
-    username = usernameParam;
-  }
-  // if (!process.env.TEST_USERNAME) {
-  //   process.env.TEST_USERNAME = username;
-  // }
-  await run(username);
-}, {});
-
-fixtures.password.init(async ({ environment, passwordParam }, run) => {
-  let password = defaultConfig.password;
-  if (environment === 'staging') {
-    password = passwordParam;
-  }
-  // if (!process.env.TEST_PASSWORD) {
-  //   process.env.TEST_PASSWORD = password;
-  // }
-  await run(password);
-}, {});
-
-fixtures.demoDeviceName.init(async ({ environment }, run) => {
-  let demoDeviceName = defaultConfig.demoDeviceName;
-  if (environment === 'staging') {
-    demoDeviceName = 'release-v1';
-  }
-  await run(demoDeviceName);
-}, {});
-
-export const { beforeAll, afterAll, beforeEach, afterEach, it, expect, describe } = fixtures.build();
+export default test;
