--- tests/e2e_tests/integration/test.spec.ts
@@ -1,8 +1,0 @@
-import './00-setup';
-import './01-login';
-import './02-layoutAssertions.ts';
-import './03-files.ts';
-import './04-deployments.ts';
-import './05-deviceDetails.ts';
-import './06-auditlogs.ts';
-import './99-settings.ts';
