+++ tests/e2e_tests/package.json
@@ -7,9 +7,9 @@
     "url": "git+https://github.com/mendersoftware/integration.git"
   },
   "scripts": {
-    "test": "folio --workers 1 --config=config.ts",
-    "test-ci": "folio --workers 1 --config=config.ts",
-    "test-visual": "PWDEBUG=1 folio --workers 1 --config=config.ts"
+    "test": "npx playwright test --workers 1 --config=config.ts",
+    "test-ci": "npx playwright test --workers 1 --config=config.ts",
+    "test-visual": "PWDEBUG=1 npx playwright test --workers 1 --config=config.ts"
   },
   "author": {
     "name": "Northern.tech"
@@ -21,7 +21,7 @@
   "homepage": "https://mender.io",
   "devDependencies": {
     "@nuintun/qrcode": "^3.0.1",
-    "@playwright/test": "0.1101",
+    "@playwright/test": "1.12.3",
     "axios": "^0.21.1",
     "dayjs": "^1.10.5",
     "jwt-decode": "^3.1.2",
